Implement hidden sets.
The easiest way to find hidden sets is to look for their associated
naked sets. In a region with n empty cells, a hidden set of size k will
always be accompanied by a naked set of size n - k.

Also fix a couple of bugs in some of the basic algorithms used.